% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyLogND.R
\name{convertLogToMean}
\alias{convertLogToMean}
\title{Convert log converted mean to the original mean}
\usage{
convertLogToMean(meanlog = 1, sdlog = 1)
}
\arguments{
\item{meanlog}{Mean of log converted distribution}

\item{sdlog}{Standard deviation of log converted distribution}
}
\value{
Mean of original distribution
}
\description{
\code{convertLogToMean} converts log converted mean to the original mean
}
