% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyLogND.R
\name{convertLogToSD}
\alias{convertLogToSD}
\title{Convert log converted SD to the original SD}
\usage{
convertLogToSD(meanlog = 1, sdlog = 1)
}
\arguments{
\item{meanlog}{Mean of log converted distribution}

\item{sdlog}{Standard deviation of log converted distribution}
}
\value{
SD of original distribution
}
\description{
\code{convertLogToSD} converts log converted SD to the original SD
}
