% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyLogND.R
\name{histLogND}
\alias{histLogND}
\title{Compare histogram between original and log converted distribution}
\usage{
histLogND(data)
}
\arguments{
\item{data}{A list of log normal distribution}
}
\description{
\code{histLogND} draws histogram of original and log converted distribution.
Red vertical line shows the mean of the original distribution.
Blue vertical line shows the mean of log converted distribution.
Green vertical line shows the median of the distribution.
}
