% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyLogND.R
\name{meanIntervalLogND}
\alias{meanIntervalLogND}
\title{Calculate interval estimate of mean with Bootstrap}
\usage{
meanIntervalLogND(data, R = 1000)
}
\arguments{
\item{data}{A list of log normal distribution}
}
\description{
\code{meanIntervalLogND} calculate interval estimate of mean.
}
