% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyLogND.R
\name{normalityCheckLogND}
\alias{normalityCheckLogND}
\title{Check normality of log converted distribution}
\usage{
normalityCheckLogND(data)
}
\arguments{
\item{data}{A list of log normal distribution}
}
\description{
\code{normalityCheckLogND} checks the normality of log converted distribution and draw QQ-plot.
}
