% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kyLogND.R
\name{rlnorm2}
\alias{rlnorm2}
\title{Generate random numbers with log normal distribution}
\usage{
rlnorm2(n, mean = 1, sd = 1)
}
\arguments{
\item{n}{Length of generated list}

\item{mean}{Mean of original distribution}

\item{sd}{Standard deviation of original distribution}
}
\value{
A list of generated random number
}
\description{
\code{rlnorm2} Generate random numbers with log normal distribution
}
